refactor: 추측 뷰와 재시작 뷰를 나누고 DTO 이름 변경
