+++ src/main/java/baseball/Application.java
@@ -4,11 +4,11 @@ import baseball.domain.RandomSecretNumberGenerator;
 import baseball.domain.SecretNumberGenerator;
 import baseball.domain.game.BaseballGame;
 import baseball.domain.game.TrialResult;
-import baseball.ui.InputView;
-import baseball.ui.OutputView;
-import baseball.ui.dto.GuessRequest;
-import baseball.ui.dto.GuessResponse;
-import baseball.ui.dto.RestartRequest;
+import baseball.ui.GuessView;
+import baseball.ui.RestartView;
+import baseball.ui.dto.GuessInput;
+import baseball.ui.dto.GuessOutput;
+import baseball.ui.dto.RestartInput;
 
 public class Application {
     public static final int SECRET_NUMBER_SIZE = 3;
@@ -24,7 +24,7 @@ public class Application {
     }
 
     private static void restartOrFinish(BaseballGame game) {
-        final RestartRequest request = InputView.inputRestart();
+        final RestartInput request = RestartView.inputRestartOrNot();
         if (request.isRestarted()) {
             game.restart();
         } else {
@@ -34,9 +34,9 @@ public class Application {
 
     private static void playOneSet(BaseballGame game) {
         while (!game.isCorrect()) {
-            final GuessRequest request = InputView.inputGuess();
+            final GuessInput request = GuessView.inputGuess();
             final TrialResult result = game.tryGuess(request.getGuessDigits());
-            OutputView.printGuessResult(new GuessResponse(result), SECRET_NUMBER_SIZE);
+            GuessView.printGuessResult(new GuessOutput(result), SECRET_NUMBER_SIZE);
         }
     }
